Move LicenseFilenamePatterns from the 'utils' to the 'model' module
This class has become part of the configuration. Thus, move it to the
dedicated module.
