+++ cli/src/main/kotlin/OrtMain.kt
@@ -42,8 +42,8 @@ import org.apache.logging.log4j.core.config.Configurator
 
 import org.ossreviewtoolkit.commands.*
 import org.ossreviewtoolkit.model.Environment
+import org.ossreviewtoolkit.model.config.LicenseFilenamePatterns
 import org.ossreviewtoolkit.model.config.OrtConfiguration
-import org.ossreviewtoolkit.utils.LicenseFilenamePatterns
 import org.ossreviewtoolkit.utils.ORT_CONFIG_DIR_ENV_NAME
 import org.ossreviewtoolkit.utils.ORT_CONFIG_FILENAME
 import org.ossreviewtoolkit.utils.ORT_DATA_DIR_ENV_NAME
