+++ model/src/main/kotlin/config/LicenseFilenamePatterns.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2020 HERE Europe B.V.
+ * Copyright (C) 2020 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,7 +17,9 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.utils
+package org.ossreviewtoolkit.model.config
+
+import org.ossreviewtoolkit.utils.log
 
 data class LicenseFilenamePatterns(
     /**
